Implement groovy_junit_test
This is very similar to spock_test, except it refers to the test classes
as "Tests" instead of "Specs", and it doesn't include the dependency on
Spock. This is the last of the Groovy rules I'm planning to contribute,
though there are a couple of improvements left to be made to the
existing rules.

--
Change-Id: Ie5acea7f65fcc96ec8cff9a3f9079583e7f77cb7
Reviewed-on: https://bazel-review.googlesource.com/#/c/2431/
MOS_MIGRATED_REVID=111694856